Refine addon imports
