+++ README.md
@@ -1,6 +1,6 @@
 # Simple Blender Addon
 
-Dieses Verzeichnis enthaelt ein minimales Blender Addon für Blender 4.4 oder neuer. Kopiere den Ordner `tracking` in deinen Blender Addons-Ordner und aktiviere das Addon in den Einstellungen.
+Dieses Verzeichnis enthaelt ein minimales Blender Addon für Blender 4.4 oder neuer. Kopiere den Ordner `t` in deinen Blender Addons-Ordner und aktiviere das Addon in den Einstellungen.
 
 Das Addon stellt einen simplen Operator bereit, der im Info-Bereich eine Meldung ausgibt.
