+++ t/operators/__init__.py
@@ -1,0 +1,4 @@
+# Absolute import to the tracking operator package
+from t.operators import tracking
+
+operator_classes = tracking.operator_classes
