+++ t/operators/tracking/__init__.py
@@ -1,0 +1,8 @@
+# Import operator modules via absolute paths
+from t.operators.tracking import solver, camera, export
+
+operator_classes = (
+    *solver.operator_classes,
+    *camera.operator_classes,
+    *export.operator_classes,
+)
