+++ t/properties/tracking_props.py
@@ -1,7 +1,7 @@
 import bpy
 from bpy.props import IntProperty, FloatProperty
 
-scene_properties = {
+tracking_properties = {
     "marker_frame": IntProperty(
         name="Marker/Frame",
         description="Frame für neuen Marker",
@@ -12,20 +12,10 @@ scene_properties = {
         description="Anzahl der Frames pro Tracking-Schritt",
         default=25,
     ),
-    "nm_count": IntProperty(
-        name="NM",
-        description="Anzahl der TEST_-Tracks nach Count",
-        default=0,
-    ),
     "threshold_value": FloatProperty(
         name="Threshold Value",
         description="Gespeicherter Threshold-Wert",
         default=1.0,
-    ),
-    "test_value": IntProperty(
-        name="Test Value",
-        description="Ergebniswert aus Testfunktionen",
-        default=0,
     ),
     "error_threshold": FloatProperty(
         name="Error Threshold",
@@ -34,11 +24,11 @@ scene_properties = {
     ),
 }
 
-def register_properties():
-    for name, prop in scene_properties.items():
+def register_props():
+    for name, prop in tracking_properties.items():
         setattr(bpy.types.Scene, name, prop)
 
-def unregister_properties():
-    for name in scene_properties.keys():
+def unregister_props():
+    for name in tracking_properties.keys():
         if hasattr(bpy.types.Scene, name):
             delattr(bpy.types.Scene, name)
