+++ t/ui/__init__.py
@@ -1,0 +1,4 @@
+# Absolute import for the panel aggregator
+from t.ui import panels as panels
+
+panel_classes = panels.panel_classes
