+++ t/ui/panels/test_panels.py
@@ -1,43 +1,5 @@
 import bpy
 
-class CLIP_PT_tracking_panel(bpy.types.Panel):
-    bl_space_type = 'CLIP_EDITOR'
-    bl_region_type = 'UI'
-    bl_category = 'Track'
-    bl_label = 'Addon Panel'
-
-    def draw(self, context):
-        layout = self.layout
-        layout.label(text="Addon Informationen")
-
-
-class CLIP_PT_final_panel(bpy.types.Panel):
-    bl_space_type = 'CLIP_EDITOR'
-    bl_region_type = 'UI'
-    bl_category = 'Addon'
-    bl_label = 'Final'
-
-    def draw(self, context):
-        layout = self.layout
-        layout.prop(context.scene, 'marker_frame', text='Marker/Frame')
-        layout.prop(context.scene, 'frames_track', text='Frames/Track')
-        layout.prop(context.scene, 'error_threshold', text='Error Threshold')
-
-
-class CLIP_PT_stufen_panel(bpy.types.Panel):
-    bl_space_type = 'CLIP_EDITOR'
-    bl_region_type = 'UI'
-    bl_category = 'Addon'
-    bl_label = 'Stufen'
-
-    def draw(self, context):
-        layout = self.layout
-        layout.operator('clip.panel_button', text='Proxy')
-        layout.operator('clip.track_nr1', text='Track Nr. 1')
-        layout.operator('clip.cleanup', text='Cleanup')
-        layout.operator('clip.track_nr2', text='Track Nr. 2')
-
-
 class CLIP_PT_test_panel(bpy.types.Panel):
     bl_space_type = 'CLIP_EDITOR'
     bl_region_type = 'UI'
@@ -89,7 +51,6 @@ class CLIP_PT_test_subpanel(bpy.types.Panel):
 
     def draw(self, context):
         layout = self.layout
-
         layout.label(text="Test Aktionen")
         layout.operator('clip.test_pattern', text='Test Pattern')
         layout.operator('clip.test_motion', text='Test Motion')
@@ -105,7 +66,6 @@ class CLIP_PT_test_detail_panel(bpy.types.Panel):
 
     def draw(self, context):
         layout = self.layout
-
         layout.operator('clip.setup_defaults', text='Test Defaults')
         layout.operator('clip.detect_button', text='Test Detect')
         layout.operator('clip.test_track', text='Test Track')
@@ -113,12 +73,9 @@ class CLIP_PT_test_detail_panel(bpy.types.Panel):
         layout.operator('clip.select_test_tracks', text='TEST select')
         layout.operator('clip.error_value', text='Error Value')
 
+
 panel_classes = (
-    CLIP_PT_tracking_panel,
-    CLIP_PT_final_panel,
-    CLIP_PT_stufen_panel,
     CLIP_PT_test_panel,
     CLIP_PT_test_subpanel,
     CLIP_PT_test_detail_panel,
 )
-
