+++ t/ui/panels/tracking_panel.py
@@ -1,0 +1,14 @@
+import bpy
+
+class CLIP_PT_tracking_panel(bpy.types.Panel):
+    bl_space_type = 'CLIP_EDITOR'
+    bl_region_type = 'UI'
+    bl_category = 'Track'
+    bl_label = 'Addon Panel'
+
+    def draw(self, context):
+        layout = self.layout
+        layout.label(text="Addon Informationen")
+
+
+panel_classes = (CLIP_PT_tracking_panel,)
